Change BE endpoints. Apply refactoring

+++ server/src/main/java/com/phoenixoft/teambalanceapp/game/service/UserGameService.java
@@ -8,8 +8,10 @@ import com.phoenixoft.teambalanceapp.game.entity.Game;
 import com.phoenixoft.teambalanceapp.game.entity.Team;
 import com.phoenixoft.teambalanceapp.game.repository.GameRepository;
 import com.phoenixoft.teambalanceapp.group.entity.Group;
-import com.phoenixoft.teambalanceapp.group.service.GroupService;
+import com.phoenixoft.teambalanceapp.group.service.UserGroupService;
+import com.phoenixoft.teambalanceapp.security.dto.CustomUser;
 import com.phoenixoft.teambalanceapp.user.entity.User;
+import com.phoenixoft.teambalanceapp.user.service.UserService;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -18,14 +20,15 @@ import java.util.List;
 
 @Service
 @AllArgsConstructor
-public class GameService {
+public class UserGameService {
 
     private final GameRepository gameRepository;
-    private final GroupService groupService;
+    private final UserService userService;
+    private final UserGroupService userGroupService;
     private final TeamBalancer teamBalancer;
 
-    public Game save(Long groupId, GameRequestDto dto) {
-        Group group = groupService.findById(groupId);
+    public Game save(Long userId, Long groupId, GameRequestDto dto) {
+        Group group = userGroupService.findGroupById(userId, groupId);
         Game newGame = new Game();
         newGame.setName(dto.getName());
         newGame.setStartDateTime(dto.getStartDateTime());
@@ -33,44 +36,38 @@ public class GameService {
         return gameRepository.save(newGame);
     }
 
-    public Game findGame(Long groupId, Long gameId) {
-        Group group = groupService.findById(groupId);
-        return group.findGame(gameId)
+    public Game findGame(Long userId, Long gameId) {
+        return userService.findById(userId).findGame(gameId)
                 .orElseThrow(() -> new ResourceNotFoundException("Game not found: " + gameId));
     }
 
-    public Game updateGame(Long groupId, Long gameId, GameRequestDto dto) {
-        Group group = groupService.findById(groupId);
-        Game gameForUpdate = group.findGame(gameId)
-                .orElseThrow(() -> new ResourceNotFoundException("Game not found: " + gameId));
+    public Game updateGame(Long userId, Long gameId, GameRequestDto dto) {
+        Game gameForUpdate = findGame(userId, gameId);
         gameForUpdate = dto.fillEntityWithDtoData(gameForUpdate);
         return gameRepository.save(gameForUpdate);
     }
 
-    public void delete(Long groupId, Long gameId) {
-        Group group = groupService.findById(groupId);
-        Game gameForDelete = group.findGame(gameId)
-                .orElseThrow(() -> new ResourceNotFoundException("Game not found: " + gameId));
-        group.removeGame(gameForDelete);
+    public void deleteGame(CustomUser user, Long gameId) {
+        Game gameForDelete = findGame(user.getId(), gameId);
+        userGroupService.checkAdminPermissions(user, gameForDelete.getGroup().getId());
         gameRepository.deleteById(gameForDelete.getId());
     }
 
-    public List<User> getGamePlayers(Long groupId, Long gameId) {
-        return findGame(groupId, gameId).getPlayers();
+    public List<User> getGamePlayers(Long userId, Long gameId) {
+        return findGame(userId, gameId).getPlayers();
     }
 
-    public Game addPlayerToGame(Long groupId, Long gameId, Long userId) {
-        Group group = groupService.findById(groupId);
-        User newPlayer = group.findMember(userId)
-                .orElseThrow(() -> new ResourceNotFoundException("Group member not found: " + userId));
-        Game game = findGame(groupId, gameId);
+    public Game addPlayerToGame(Long userId, Long gameId, Long memberToAddId) {
+        Game game = findGame(userId, gameId);
+        User newPlayer = game.getGroup().findMember(memberToAddId)
+                .orElseThrow(() -> new ResourceNotFoundException("Group member not found: " + memberToAddId));
         game.getPlayers().add(newPlayer);
         return gameRepository.save(game);
     }
 
-    public Game addPlayersToGame(Long groupId, Long gameId, AddPlayersRequestDto addPlayersRequestDto) {
-        Group group = groupService.findById(groupId);
-        Game game = findGame(groupId, gameId);
+    public Game addPlayersToGame(Long userId, Long gameId, AddPlayersRequestDto addPlayersRequestDto) {
+        Game game = findGame(userId, gameId);
+        Group group = game.getGroup();
         for (Long playerId : addPlayersRequestDto.getPlayers()) {
             User newPlayer = group.findMember(playerId)
                     .orElseThrow(() -> new ResourceNotFoundException("Group member not found: " + playerId));
@@ -79,17 +76,17 @@ public class GameService {
         return gameRepository.save(game);
     }
 
-    public Game deletePlayerFromGame(Long groupId, Long gameId, Long userId) {
-        Game game = findGame(groupId, gameId);
-        if (game.removePlayer(userId)) {
+    public Game deletePlayerFromGame(Long userId, Long gameId, Long playerId) {
+        Game game = findGame(userId, gameId);
+        if (game.removePlayer(playerId)) {
             game.setBalancedTeams(null);
             game = gameRepository.save(game);
         }
         return game;
     }
 
-    public List<Team> generateBalancedTeams(Long groupId, Long gameId, int teamsCount) {
-        Game game = findGame(groupId, gameId);
+    public List<Team> generateBalancedTeams(Long userId, Long gameId, int teamsCount) {
+        Game game = findGame(userId, gameId);
         List<Team> teams = game.getBalancedTeams() == null
                 ? teamBalancer.dividePlayersIntoBalancedTeams(new ArrayList<>(game.getPlayers()), teamsCount)
                 : teamBalancer.dividePlayersIntoBalancedTeamsWithSomeRandomness(new ArrayList<>(game.getPlayers()), teamsCount);
